chatbot/chatbot/requirements.txt
gradio
transformers
torch
sentencepiece
soundfile
openai-whisper
gTTS
numpy
scipy
ffmpeg-python
fastapi 
uvicorn
